third_party/WebKit/LayoutTests/platform/win/virtual/mojo-loading/http/tests/navigation/pushstate-whitelisted-denied-expected.txt

This is a testharness.js-based test.
PASS testRunner.addOriginAccessWhitelistEntry is required for this test 
PASS pushState to origin with different port should fail with SecurityError (even with whitelisted origins) 
Harness: the test ran to completion.
